parametrized split tests to remove nesting and improve pylint score, changed import in splitting factory to fix import problem after folder restructure
